An online portal for doctors,laboraroties and patients
The “SMART TREATMENT” is a system which brings Patients, Doctors and Laboratories together.
In the current system, when a patient is suffering from some disease, the patient goes to a hospital or to a private practitioner who he can know earlier. The system to consult a private practitioner and laboratories is still manual. There is no automated system for consulting private practitioners and laboratories. The proposed system solves this problem by making the entire existing system into an online application.
Using the “SMART TREATMENT” system  when a patient suffering from any disease gives their place and needed specialist as input to the system. Patient can give request to doctor. The doctor will give the appointment time and place to the patient. If checkup is needed the patient can give request to access the laboratories. The proposed system also allows the doctor to suggest laboratories to the patient. The laboratory will give the appointment time to the patient. Then the patient goes to laboratory and gives input (blood,x-ray,etc). After paying fees to the laboratories the result is uploaded by them to the system and it can be accessed by both the patient and doctor. No need to weight for a hard copy result. Then the patient can obtain prescription from private practitioner easily.